requirements.txt
coverage==4.5.4
Django==3.0.5
djangorestframework==3.11.0
Inject==4.1.2
python-coveralls==2.9.3
python-magic==0.4.15
# asgiref==3.2.7
# astroid==2.3.3
# certifi==2020.4.5.1
# chardet==3.0.4
# idna==2.9
# isort==4.3.21
# lazy-object-proxy==1.4.3
# mccabe==0.6.1
# pylint==2.4.4
# pytz==2019.3
# PyYAML==5.3.1
# requests==2.23.0
# six==1.14.0
# sqlparse==0.3.1
# urllib3==1.25.9
# wrapt==1.11.2
